backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.1
supabase==2.6.0
flasgger==0.9.7.1